pyproject.toml
[project]
name = "mcp-mitm-mem0"
version = "0.2.0"
description = "Memory service for Claude via MITM proxy and MCP using Mem0 SaaS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.4",
    "mem0ai>=1.0.0",
    "mitmproxy>=12.1.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
]

[project.scripts]
mcp-mitm-mem0 = "mcp_mitm_mem0.mcp_server:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]
lint = [
    "ruff>=0.11.0",
]

[tool.ruff]
preview = true
fix = true
unsafe-fixes = true
cache-dir = "~/.cache/ruff"
target-version = "py312"
output-format = "grouped"

[tool.ruff.lint]
preview = true
select = ["E4", "E7", "E9", "F", "B", "C4", "SIM", "UP", "S", "N", "I", "Q"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
preview = true

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Enable auto-formatting of code examples in docstrings.
docstring-code-format = true